Revert "networkd: netdev - set predictable mac address when creating netdev"
This reverts commit cdc85c875b842b9309f72caefc51c262f521cf92.

There appears to be a kernel bug that (among other things) cause bridges
not to get a bridge id set when supplying IFLA_ADDRESS when creating the
netdev.

Simply revert the whole thing until we sort this out in the kernel.

See: http://www.spinics.net/lists/netdev/msg279807.html

Reported-by: C. R. Oldham <cr@saltstack.com>