path-util: add path_make_relative()
In user_dirs() in path-lookup.c, I want to replace this:
        symlink("../../../.config/systemd/user", data_home);
with
        symlink(config_home, data_home);
to avoid hardcoding .config when XDG_CONFIG_HOME is set.

The problem is that config_home is an absolute path, and it's better
to make the symlink relative. path_make_relative() is an utility
function that converts an absolute path into a relative one.